refactor: 지역 선택 데이터를 data/regions.ts로 폴더명 변경 및 수정 #13
