+++ eatmoji/src/data/regions.ts
